+++ agents/management/pyproject.toml
@@ -9,4 +9,7 @@ version = "0.1.0"
 description = "엔터테인먼트 컨텐츠 관리를 위한 LangGraph Workflow 모듈"
 readme = "README.md"
 requires-python = ">=3.13"
-dependencies = []
+dependencies = [
+    "langchain-google-genai>=2.1.8",
+    "pandas>=2.3.1",
+]
